day5

+++ day.hpp
@@ -48,11 +48,11 @@ class Day {
 
       auto pos = s.find(delim);
       if (pos == std::string::npos) {
-        numbers.emplace_back(std::stoi(s));
+        numbers.emplace_back(std::stoll(s));
         break;
       }
 
-      numbers.emplace_back(std::stoi(s.substr(0, pos)));
+      numbers.emplace_back(std::stoll(s.substr(0, pos)));
       s = s.erase(0, pos);
     }
